Moving method from DotTemplateTool to UT
